feat: restructure services into subdirectories and implement new service files
