+++ Apps/client/src/services/maintenance/maintenanceStrategy.ts
@@ -1,5 +1,5 @@
 import { AlgorithmResponse, MaintenancePlanningResultDto } from "@/types";
-import { apiFetch } from "./api";
+import { apiFetch } from "../api";
 
 export async function getMaintenancePlan(budget: number): Promise<AlgorithmResponse<MaintenancePlanningResultDto>> {
   const params = new URLSearchParams({ budget: budget.toString() });
