+++ Apps/client/src/services/network/networkExpansion.ts
@@ -1,5 +1,5 @@
 import { AlgorithmResponse, MstResultDto } from "@/types";
-import { apiFetch } from "./api";
+import { apiFetch } from "../api";
 
 export async function getCheapestNetwork(): Promise<AlgorithmResponse<MstResultDto>> {
   return apiFetch<AlgorithmResponse<MstResultDto>>("network-expansion");
