+++ Apps/client/src/services/network/networkTopology.ts
@@ -1,5 +1,5 @@
 import {  NetworkTopologyData } from "@/types";
-import { apiFetch } from "./api";
+import { apiFetch } from "../api";
 
 export async function fetchNetworkTopology(): Promise<NetworkTopologyData> {
   return apiFetch<NetworkTopologyData>("network-topology");
