+++ Apps/client/src/services/routes/emergencyRouting.ts
@@ -1,5 +1,5 @@
 import { AlgorithmResponse, ShortestPathResultDto } from "@/types";
-import { apiFetch } from "./api";
+import { apiFetch } from "../api";
 
 export async function getEmergencyRoute(from: string, to: string): Promise<AlgorithmResponse<ShortestPathResultDto>> {
   const params = new URLSearchParams({ from, to });
